better docker_project_install

+++ libs/apps/nginx_helper.sh
@@ -30,8 +30,6 @@ function nginx_server_create() {
     local redirect_domains="${4}"
     local proxy_port="${5}"
 
-    #local debug
-
     # Log
     log_event "debug" "Project type: ${project_type}" "false"
     log_event "debug" "Server type: ${server_type}" "false"
@@ -92,12 +90,8 @@ function nginx_server_create() {
         cp "${BROLIT_MAIN_DIR}/config/nginx/sites-available/${project_type}_${server_type}" "${nginx_server_file}"
 
         # -L returns true if the "file" exists and is a symbolic link
-        if [[ -L ${nginx_server_file_link} ]]; then
-
-            # Remove previous symbolic link
-            rm "${nginx_server_file_link}"
-
-        fi
+        ## Remove previous symbolic link
+        [[ -L ${nginx_server_file_link} ]] && rm "${nginx_server_file_link}"
 
         # Creating symbolic link
         ln -s "${nginx_server_file}" "${nginx_server_file_link}"
@@ -213,15 +207,14 @@ function nginx_server_change_status() {
         log_event "info" "New project status: ${project_status}" "false"
 
         if [[ -f "${WSERVER}/sites-available/${project_domain}" ]]; then
-
             # Creating symbolic link
             ln -s "${WSERVER}/sites-available/${project_domain}" "${WSERVER}/sites-enabled/${project_domain}"
-            # Logging
+            # Log
             log_event "info" "Project config added to ${WSERVER}/sites-enabled/${project_domain}" "false"
             display --indent 6 --text "- Changing project status to ONLINE" --result "DONE" --color GREEN
 
         else
-            # Logging
+            # Log
             log_event "error" "${WSERVER}/sites-available/${project_domain} does not exist" "false"
             display --indent 6 --text "- Changing project status to ONLINE" --result "FAIL" --color RED
             display --indent 8 --text "${WSERVER}/sites-available/${project_domain} does not exist" --tcolor RED
@@ -322,7 +315,7 @@ function nginx_server_get_current_phpv() {
     local nginx_server_file="${1}"
 
     # Replace string to match PHP version
-    current_php_v_string=$(cat ${nginx_server_file} | grep fastcgi_pass | cut -d '/' -f 4 | cut -d '-' -f 1)
+    current_php_v_string=$(cat "${nginx_server_file}" | grep fastcgi_pass | cut -d '/' -f 4 | cut -d '-' -f 1)
     current_php_v=${current_php_v_string#"php"}
 
     # Log
@@ -602,15 +595,17 @@ function nginx_generate_encrypted_auth() {
     printf "${user}:${encrypted_psw}" >"/etc/nginx/.passwords"
 
     exitstatus=$?
-    if [[ ${exitstatus} -eq 1 ]]; then
+    if [[ ${exitstatus} -eq 0 ]]; then
+
+        chmod 640 "/etc/nginx/.passwords"
+        chown www-data:www-data "/etc/nginx/.passwords"
+
+    else
 
         log_event "error" "Something went wrong writing: /etc/nginx/.passwords" "false"
         return 1
 
     fi
-
-    chmod 640 "/etc/nginx/.passwords"
-    chown www-data:www-data "/etc/nginx/.passwords"
 
 }
 
@@ -631,12 +626,8 @@ function nginx_server_add_http2_support() {
     # Check if the file exists
     nginx_server_file="/etc/nginx/sites-available/${nginx_server_file}"
 
-    if [[ ! -f "${nginx_server_file}" ]]; then
-
-        log_event "error" "File ${nginx_server_file} not found" "false"
-        return 1
-
-    fi
+    # Return error if file doesn't exist
+    [[ ! -f "${nginx_server_file}" ]] && log_event "error" "File ${nginx_server_file} not found" "false" && return 1
 
     # Add http2 to ports
     sed -i "s/listen 443 ssl;/listen 443 ssl http2;/g" "${nginx_server_file}"
